improve testability by sorting tests into files that mirror lib + group by method
